refactor: LottoAutoService 책임을 LottoWinService 클래스로 분리  - LottoAutoService 의 당첨조회 기능을 LottoWinService 클래스로 이동  - LottoWinInquiryController 클래스 의존관계 수정  - 테스트코드 작성
