+++ src/test/java/lotto/service/LottoAutoServiceTest.java
@@ -1,15 +1,10 @@
 package lotto.service;
 
-import lotto.domain.LottoRank;
-import lotto.domain.LottoRanks;
 import lotto.domain.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -55,55 +50,6 @@ class LottoAutoServiceTest {
 
         // then
         assertThat(lottoTicket.isAutoTicket()).isTrue();
-    }
-
-    @Test
-    @DisplayName("inquiryWin 기본테스트")
-    void inquiryWin() {
-        // given
-        List<String> lottoNumberInputList = new ArrayList<>();
-        lottoNumberInputList.add("1, 2, 3, 4, 5, 6");
-        LottoAutoService lottoAutoService = new LottoAutoService();
-
-        // when
-        LottoRanks lottoRanks = lottoAutoService.inquiryWin(lottoNumberInputList, new int[]{1, 2, 3, 4, 5, 6}, 45);
-
-        // then
-        assertThat(1).isEqualTo(lottoRanks.matchLottoCount(LottoRank.FIRST));
-    }
-
-    @Test
-    @DisplayName("수익률 계산")
-    void totalReturnRate() {
-        // given
-        List<String> lottoNumberInputList = new ArrayList<>();
-        lottoNumberInputList.add("1, 2, 3, 4, 5, 6");
-        lottoNumberInputList.add("1, 2, 3, 4, 5, 7");
-        lottoNumberInputList.add("1, 2, 3, 4, 7, 8");
-        LottoAutoService lottoAutoService = new LottoAutoService();
-
-        // when
-        LottoRanks lottoRanks = lottoAutoService.inquiryWin(lottoNumberInputList, new int[]{1, 2, 3, 4, 5, 6}, 45);
-
-        // then
-        assertThat(667183.33).isEqualTo(lottoRanks.totalReturnRate());
-    }
-
-    @Test
-    @DisplayName("수익률 계산 - 보너스번호")
-    void totalReturnRate_withBonusNumber() {
-        // given
-        List<String> lottoNumberInputList = new ArrayList<>();
-        lottoNumberInputList.add("1, 2, 3, 4, 5, 6");
-        lottoNumberInputList.add("1, 2, 3, 4, 5, 7");
-        lottoNumberInputList.add("1, 2, 3, 4, 7, 8");
-        LottoAutoService lottoAutoService = new LottoAutoService();
-
-        // when
-        LottoRanks lottoRanks = lottoAutoService.inquiryWin(lottoNumberInputList, new int[]{1, 2, 3, 4, 5, 6}, 7);
-
-        // then
-        assertThat(676683.33).isEqualTo(lottoRanks.totalReturnRate());
     }
 
 }
